Added grid operators separately, including convective_derivative
