+++ src/ImmersedLayers.jl
@@ -20,12 +20,13 @@ export DoubleLayer, SingleLayer, MaskType, Mask, ComplementaryMask,
         SurfaceCache,SurfaceScalarCache,SurfaceVectorCache,
         BasicScalarILMProblem,BasicVectorILMProblem,prob_cache,
         AbstractScalingType,GridScaling,IndexScaling,
-        BasicILMCache,
+        BasicILMCache,ILMSystem,
         AbstractExtraILMCache,AbstractScalarILMProblem,AbstractVectorILMProblem,
+        ConvectiveDerivativeCache,convective_derivative, convective_derivative!,
         regularize_normal!,normal_interpolate!,
         surface_curl!,surface_divergence!,surface_grad!,inverse_laplacian!,
         mask,mask!,complementary_mask,complementary_mask!,
-        CLinvCT,GLinvD,nRTRn
+        create_CLinvCT,create_GLinvD,create_nRTRn
 
 abstract type LayerType{N} end
 
@@ -34,7 +35,9 @@ include("cache.jl")
 include("problem.jl")
 include("system.jl")
 include("layers.jl")
+include("grid_operators.jl")
 include("surface_operators.jl")
+include("matrix_operators.jl")
