+++ src/surface_operators.jl
@@ -186,23 +186,6 @@ function _unscaled_surface_grad!(vn::VectorData{N},ϕ::Edges{Primal,NX,NY},nrm::
 end
 
 """
-    inverse_laplacian!(w::GridData,sys::ILMSystem)
-
-Compute the in-place inverse Laplacian of grid data `w`, and multiply the result
-by unity or by the grid cell size, depending on whether `sys` has `IndexScaling` or `GridScaling`,
-respectively.
-"""
-inverse_laplacian!(w,sys::ILMSystem) = inverse_laplacian!(w,sys.base_cache)
-
-function inverse_laplacian!(w::GridData,cache::BasicILMCache)
-    @unpack L = cache
-    _unscaled_inverse_laplacian!(w,L)
-    _scale_inverse_laplacian!(w,cache)
-end
-
-_unscaled_inverse_laplacian!(w::GridData,L::CartesianGrids.Laplacian) = w .= L\w
-
-"""
     mask(sys::ILMSystem) -> GridData
 
 Create grid data that consist of 1s inside of a surface (i.e., on a side opposite
@@ -261,100 +244,3 @@ function _complementary_mask!(msk,cache)
     _mask!(msk,cache)
     msk .= 1.0 - msk
 end
-
-"""
-    CLinvCT(cache::BasicILMCache[;scale=1.0])
-
-Construct the square matrix ``-C_s L^{-1}C_s^T``, which maps data of type `ScalarData`
-to data of the same type. The operators `C_s` and `C_s^T` correspond to `surface_curl!`
-and `L` is the grid Laplacian. The optional keyword `scale` multiplies the
-matrix by the designated value.
-"""
-function CLinvCT(cache::BasicILMCache{N};scale=1.0) where {N}
-    @unpack L, sdata_cache, gcurl_cache = cache
-
-    len = length(sdata_cache)
-    A = Matrix{eltype(sdata_cache)}(undef,len,len)
-    fill!(sdata_cache,0.0)
-
-    for col in 1:len
-        sdata_cache[col] = 1.0
-        fill!(gcurl_cache,0.0)
-        surface_curl!(gcurl_cache,sdata_cache,cache)
-
-        inverse_laplacian!(gcurl_cache,cache)
-        gcurl_cache .*= -1
-        surface_curl!(sdata_cache,gcurl_cache,cache)
-
-        A[:,col] = scale*sdata_cache
-        fill!(sdata_cache,0.0)
-    end
-
-    return A
-
-end
-
-"""
-    GLinvD(cache::BasicILMCache[;scale=1.0])
-
-Construct the square matrix ``G_s L^{-1}D_s``, which maps data of type `ScalarData`
-to data of the same type. The operators `G_s` and `D_s` correspond to `surface_grad!`
-and `surface_divergence!`, and `L` is the grid Laplacian. The optional keyword `scale` multiplies the
-matrix by the designated value.
-"""
-function GLinvD(cache::BasicILMCache{N};scale=1.0) where {N}
-    @unpack L, sdata_cache, gdata_cache = cache
-
-    len = length(sdata_cache)
-    A = Matrix{eltype(sdata_cache)}(undef,len,len)
-    fill!(sdata_cache,0.0)
-
-    for col in 1:len
-        sdata_cache[col] = 1.0
-        fill!(gdata_cache,0.0)
-        surface_divergence!(gdata_cache,sdata_cache,cache)
-        inverse_laplacian!(gdata_cache,cache)
-        surface_grad!(sdata_cache,gdata_cache,cache)
-
-        A[:,col] .= scale*sdata_cache
-        fill!(sdata_cache,0.0)
-    end
-
-    return A
-
-end
-
-"""
-    nRTRn(cache::BasicILMCache[;scale=1.0])
-
-Construct the square matrix ``n\\cdot I_f R_f n \\circ``, which maps data of type `ScalarData`
-to data of the same type. The operators `I_f` and `R_f` correspond to the interpolation
-and regularization matrices. The optional keyword `scale` multiplies the
-matrix by the designated value.
-"""
-function nRTRn(cache::BasicILMCache{N};scale=1.0) where {N}
-    @unpack sdata_cache, gsnorm_cache = cache
-
-    len = length(sdata_cache)
-    A = Matrix{eltype(sdata_cache)}(undef,len,len)
-    fill!(sdata_cache,0.0)
-
-    for col in 1:len
-        sdata_cache[col] = 1.0
-        fill!(gsnorm_cache,0.0)
-        regularize_normal!(gsnorm_cache,sdata_cache,cache)
-        normal_interpolate!(sdata_cache,gsnorm_cache,cache)
-
-        A[:,col] .= scale*sdata_cache
-        fill!(sdata_cache,0.0)
-    end
-
-    return A
-
-end
-
-_scale_derivative!(w,cache::BasicILMCache{N,IndexScaling}) where {N} = w
-_scale_derivative!(w,cache::BasicILMCache{N,GridScaling}) where {N} = w ./= cellsize(cache)
-
-_scale_inverse_laplacian!(w,cache::BasicILMCache{N,IndexScaling}) where {N} = w
-_scale_inverse_laplacian!(w,cache::BasicILMCache{N,GridScaling}) where {N} = w .*= cellsize(cache)^2
